removing warning message for future op, since this is now expected; debounce getLatestOps
with the new server-side query subscriptions, getLatestOps is expected
to be called whenever a query newly matches a doc. doing it this way
and relying on the client to send a request for the ops it is missing
or the full document eliminates needing to worry about much server
state, and the server can just blindly start piping through ops as soon
as the query results change. Needed to implement debouncing, since this
is an expected normal mechanism now and a bunch of ops might get sent
before the client has a chance to catch up

Conflicts:
	lib/client/doc.js